Update: use later version of emoji-data.txt
I also had to tweak the code that read the part of the comment that
contains the Unicode version when a glyph was introduced. Some entries in
the newer version of the emoji-data.txt file are for reserved CodePoints
and the version of "NA" was breaking on code that was expecting a floating
point number. For those we return a version of 0.0 instead and then
ignore them.

I also had to commemnt out the test code in the Makefile as the
alternative that was being tested against was using Unicode 9.0 datai
and is not longer comparible as this is now set for Unicode 12 and
should be compatible with later version when they emerge.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>