BugFix: change out wchar_t type which is too narrow on Windows
Specifically it is only 16 bits and as we are using it to pass a Unicode
codepoint it needs to have at least 21! This does mean however that any
surrogate pairs i.e. UTF-16 encoded characters that are not on the BMP
and are conveyed by TWO 16 bit values, MUST be converted to a UTF-32 value
before being fed to `(int) widechar_wcwidth(int32_t c)`!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>