Add Rust module output
This commit adds a template for a Rust module that provides equivalent
functionality to the C++ implementation.

`make rust` will verify the syntax and run the handful of very basic
tests; since the tables and logic are the same as those used by the C++
implementation, the Rust tests are really just there to sniff test that
the binary search is working for a few inputs.