Fix error on empty code point range
merged_codepoints() first checks if it's input is truthy (to return
on empty inputs), and then sorts it, converting it to a list. This
breaks if the input is an empty iterator, which is truthy, but
will produce an empty list, which we want to ignore.

Fix this by converting to a list first. While at it, add a type
annotation.

	$ make
	./generate.py
	Thinking...
	Traceback (most recent call last):
	  File "/home/johannes/git/widecharwidth/./generate.py", line 742, in <module>
	    fields = make_fields(datas, cps, LangSettings("{}"), OUTPUT_FILENAME)
	  File "/home/johannes/git/widecharwidth/./generate.py", line 734, in make_fields
	    "widenedin9": codepoints_with_width(WIDTH_WIDENED_IN_9),
	  File "/home/johannes/git/widecharwidth/./generate.py", line 709, in codepoints_with_width
	    return codepoints_to_carray_str(
	  File "/home/johannes/git/widecharwidth/./generate.py", line 538, in codepoints_to_carray_str
	    ranges = merged_codepoints(cps)
	  File "/home/johannes/git/widecharwidth/./generate.py", line 513, in merged_codepoints
	    ranges = [(cps[0], cps[0])]
	IndexError: list index out of range