Added a pure C template.
The implementation of widechar_in_table() is similar to wcwidth9()'s, but was
written independently of it.  They share no code.

The widechar_in_table() function does a binary search on the supplied
widechar_range array, and depends on that data remaining pre-sorted.

The test.cpp code has a commented out block for manually switching to the C
implementation of widechar_width().  The tester passes when the C
implementation is used.  A separate test program for the C implementation could
be added later, but was left as a manual change for now.